Rename nrn_soa to mem_layout_util
