+++ coreneuron/io/mem_layout_util.hpp
